validate ownership and permissions of tmp dir
Fix EBPF-1183

+++ pkg/util/kernel/headers/find_headers.go
@@ -23,6 +23,7 @@ import (
 	"slices"
 	"strconv"
 	"sync"
+	"syscall"
 
 	"github.com/DataDog/datadog-agent/comp/core/telemetry/def"
 	telemetryimpl "github.com/DataDog/datadog-agent/comp/core/telemetry/impl"
@@ -200,7 +201,7 @@ func (h *headerProvider) getKernelHeaders(hv kernel.Version) ([]string, headerFe
 		// The `kheaders` module will be automatically added and removed if present and needed.
 		var err error
 		var dirs []string
-		if dirs, err = getSysfsHeaderDirs(hv); err == nil {
+		if dirs, err = getSysfsHeaderDirs(os.TempDir(), hv); err == nil {
 			return dirs, sysfsHeadersFound, nil
 		}
 		log.Debugf("unable to find system kernel headers: %v", err)
@@ -390,23 +391,50 @@ func getDownloadedHeaderDirs(headerDownloadDir string) []string {
 	return dirs
 }
 
-func getSysfsHeaderDirs(v kernel.Version) ([]string, error) {
-	tmpPath := filepath.Join(os.TempDir(), fmt.Sprintf("linux-headers-%s", v))
-	fi, err := os.Stat(tmpPath)
-	if err == nil && fi.IsDir() {
-		hv, err := getHeaderVersion(tmpPath)
-		if err != nil {
-			// remove tmp dir if it errors
-			_ = os.RemoveAll(tmpPath)
-			return nil, fmt.Errorf("unable to verify headers version: %w", err)
+var errInvalidTempDirectory = errors.New("invalid system-probe temp directory")
+
+func getSysfsHeaderDirs(tmpDir string, v kernel.Version) ([]string, error) {
+	sysprobeTmpPath := filepath.Join(tmpDir, "system-probe")
+	tmpPath := filepath.Join(sysprobeTmpPath, fmt.Sprintf("linux-headers-%s", v))
+
+	sfi, statErr := os.Stat(sysprobeTmpPath)
+	if statErr != nil {
+		if errors.Is(statErr, fs.ErrNotExist) {
+			if err := os.MkdirAll(sysprobeTmpPath, 0755); err != nil {
+				return nil, fmt.Errorf("create system-probe temp directory: %s", err)
+			}
+		} else {
+			return nil, fmt.Errorf("stat %s: %s", sysprobeTmpPath, statErr)
 		}
-		if hv != v {
-			// remove tmp dir if it fails to validate
-			_ = os.RemoveAll(tmpPath)
-			return nil, fmt.Errorf("header version %s does not match expected host version %s", v, hv)
+	} else {
+		if !sfi.IsDir() {
+			return nil, fmt.Errorf("%w: %s is not a directory", errInvalidTempDirectory, sysprobeTmpPath)
 		}
-		log.Debugf("found valid kernel headers at %s", tmpPath)
-		return []string{tmpPath}, nil
+
+		stat, ok := sfi.Sys().(*syscall.Stat_t)
+		if !ok {
+			return nil, fmt.Errorf("unable to verify permissions of %s", sysprobeTmpPath)
+		}
+		if stat.Uid != 0 || stat.Gid != 0 || sfi.Mode().Perm()&os.FileMode(0022) != 0 {
+			return nil, fmt.Errorf("%w: %s has incorrect permissions: user=%d, group=%d, permissions=%s", errInvalidTempDirectory, sysprobeTmpPath, stat.Uid, stat.Gid, sfi.Mode().Perm())
+		}
+
+		fi, err := os.Stat(tmpPath)
+		if err == nil && fi.IsDir() {
+			hv, err := getHeaderVersion(tmpPath)
+			if err != nil {
+				// remove tmp dir if it errors
+				_ = os.RemoveAll(tmpPath)
+				return nil, fmt.Errorf("unable to verify headers version: %w", err)
+			}
+			if hv != v {
+				// remove tmp dir if it fails to validate
+				_ = os.RemoveAll(tmpPath)
+				return nil, fmt.Errorf("header version %s does not match expected host version %s", v, hv)
+			}
+			log.Debugf("found valid kernel headers at %s", tmpPath)
+			return []string{tmpPath}, nil
+		}
 	}
 
 	if !sysfsHeadersExist() {
@@ -419,7 +447,7 @@ func getSysfsHeaderDirs(v kernel.Version) ([]string, error) {
 		}
 	}
 
-	if err = archive.TarXZExtractAll(sysfsHeadersPath, tmpPath); err != nil {
+	if err := archive.TarXZExtractAll(sysfsHeadersPath, tmpPath); err != nil {
 		return nil, fmt.Errorf("unable to extract kernel headers: %w", err)
 	}
 	log.Debugf("found valid kernel headers at %s", tmpPath)
